ModelOrderReductionToolkit.jl Docstrings

Projections functionality:

ModelOrderReductionToolkit.singular_values_information — Function

singular_values_information(σs)

Given a vector of real singular values σs, return a vector providing the maximum percentage of "information" available from a d-dimensional representation in Frobenius norm.

Uses the fact that the maximum error in projecting onto the first d left singular vectors is given by

||A-A_d||_F = √ (σ_{d+1}^2 + ... + σ_N^2).

So the information with a d-dimensional representation given by

singular_values_information[d] = 1 - √ (σ_{d+1}^2 + ... + σ_N^2) / √ (σ_1^2 + ... + σ_N^2)

ModelOrderReductionToolkit.pca_projector — Function

P = pca_projector(A, dim=-1; info_tol=1e-2)

Uses the first dim singular vectors of A to generate an orthogonal projection operator P. Operator performs matrix operation and can be done on vectors or matrices of proper dimension.

If dim is not provided, uses n singular values where n is the smallest integer such that the amount of information per the Frobenious norm is at least 1-info_tol.

Returns a functor P which stores the reduced dimension n, the dimension of vectors to project m, the singular values of A in s, and the vectors to project onto in the m × n matrix M.

Call with P(b) where b is a m vector or a matrix with m rows, projecting each column.

ModelOrderReductionToolkit.qr_projector — Function

qr_projector(A, dim)

Uses the first dim columns of A by qr-pivot order (norm) to generate an orthogonal projection operator P. Operator performs matrix operation and can be done on vectors or matrices of proper dimension.

Returns a functor P which stores the reduced dimension n, the dimension of vectors to project m, the vectors to project onto in the m × n matrix M, and the matrix inv(M'M) in MtMinv.

Call with P(b) where b is a m vector or a matrix with m rows, projecting each column.

ModelOrderReductionToolkit.full_lu — Function

full_lu(A; steps=-1)

Performs a completely pivoted LU factorization on the matrix A, returning permutation vectors Q and P, and lower and upper triangular matrices L and U, such that if all steps are performed, then A[P,Q] = L*U.

ModelOrderReductionToolkit.eim_projector — Function

P = eim_projector(A, dim)

Uses discrete empirical interpolation to generate a projector operator P that selects a few rows from the given vector or matrix, and then projects them onto a set of columns chosen by a greedy procedure.

Returns a functor P which stores the reduced dimension n, the dimension of vectors to project m, the matrix A_C which is A with n extracted columns, and the matrix A_CR which is A with n extracted rows and columns.

Call with P(b) where b is a m vector or a matrix with m rows, projecting each column.

Greedy linear affine reduced basis

ModelOrderReductionToolkit.Greedy_RB_Affine_Linear — Type

Greedy_RB_Affine_Linear

Struct for containing initialized greedy reduced-basis method for parametrized problem A(p) u = b(p) with affine parameter dependence A(p) = ∑ makeθAi(p,i) Ais[i], and b(p) = ∑ makeθbi(p,i) bis[i].

Uses Galerkin projection onto span of columns of V, V' A(p) V u_r = V' b(p), with V u_r ≈ u = A(p)^(-1) b(p).

Given a new parameter vector, p, and an object greedy_sol::Greedy_RB_Affine_Linear, form the reduced basis solution with greedy_sol(p[, full=true]).

ModelOrderReductionToolkit.GreedyRBAffineLinear — Function

GreedyRBAffineLinear(param_disc, Ais, makeθAi, bis, makeθbi, approx_stability_factor[, ϵ=1e-2, param_disc=nothing; max_snapshots=-1, noise=1])

Constructs a Greedy_RB_Affine_Linear object for parametrized problem A(p) y(p) = b(p) with affine parameter dependence: A(p) = ∑ makeθAi(p,i) Ais[i], and b(p) = ∑ makeθbi(p,i) bis[i].

param_disc must either be a matrix with columns as parameters, or a vector of parameter vectors. Uses the function approx_stability_factor to approximate the stability factor for each parameter in the discretization, chooses first parameter to be one with smallest stability factor. Afterwards, utilizes the affine decomposition of A(p) and b(p) to compute the norm of the residual, and compute upper-bounds for the error: ||u - V u_r|| <= ||b(p) - A(p) V u_r|| / σ_min(A(p)).

Greedily chooses parameters and then forms truth solution for parameter that results in largest upper-bound for error. Then computes ||u - V u_r||, and loops, adding to reduced basis, V, until the truth error is less than ϵ.

If max_snapshots specified, will halt after that many if ϵ accuracy not yet reached. noise determines amount of printed output, 0 for no output, 1 for basic, and 2 for more.

ModelOrderReductionToolkit.greedy_rb_err_data — Function

greedy_rb_err_data(param_disc,Ais,makeθAi,bis,makeθbi,approx_stability_factor[,num_snapshots=10,ϵ=1e-2;noise=1)

Generates error data for parametrized problem A(p) u(p) = b(p) with affine parameter dependence: A(p) = ∑ makeθAi(p,i) Ais[i], and b(p) = ∑ makeθbi(p,i) bis[i].

Note: This method calls the full order solver on all parameters in param_disc, may take a long time to run.

See the docs for GreedyRBAffineLinear for more information.

Returns a dictionary, ret_data, with the following components:

ret_data[:basis_dim] - A vector of reduced basis dimensions from 2 to num_snapshots

ret_data[:weak_greedy_ub] - A vector of the maximum upperbound l2 error found by the (weak) greedy reduced basis method (see GreedyRBAffineLinear)

ret_data[:weak_greedy_true] - A vector of the truth l2 error of the vector chosen by the (weak) greedy reduced basis method (see GreedyRBAffineLinear)

ret_data[:weak_greedy_true_ub] - A vector of the maximum true l2 error found by the (weak) greedy reduced basis method (see GreedyRBAffineLinear)

ret_data[:strong_greedy_err] - A vector of the maximum l2 error found by a strong greedy reduced basis method, uses knowledge of all solutions

ret_data[:strong_greedy_proj] - A vector of the maximum l2 error found by projecting truth solutions onto the strong greedy reduced basis, uses knowledge of all solutions

ret_data[:pca_err] - A vector of the maximum l2 error found by a PCA/POD reduced basis method, uses knowledge of all solutions

ret_data[:strong_greedy_proj] - A vector of the maximum l2 error found by projecting truth solutions onto the PCA/POD reduced basis, uses knowledge of all solutions

ModelOrderReductionToolkit.append_affine_rbm! — Function

append_affine_rbm!(x, res_init, V, Ais, VtAVis, bis, Vtbis)

Given a new vector x to append as a new column to V, update the matrices in VtAVis and the vectors in Vtbis. Also, return VtAV and Vtb which are preallocated, properly sized, matrices. Must pass in type T.

ModelOrderReductionToolkit.init_affine_rbm — Function

init_affine_rbm(x, Ais, bis, makeθAi, makeθbi, T)

Given a vector x, generate the matrices V, VtAVi, and VtAVis, and the vectors Vtb and Vtbis for preallocation and quick computation of reduced basis solutions for the problem A(p)x(p)=b(p) with affinely dependent matrix A(p) = ∑ makeθAi(p,i) Ais[i] and vector b(p) = ∑ makeθbi(p,i) bis[i]. Must pass in type T.

Successive constraint method (SCM):

ModelOrderReductionToolkit.SCM_Init — Type

SCM_Init is a struct for holding all of the variables and methods necessary for running the successive constraint method on an affinely-parameter-dependent matrix A(p) = ∑_{i=1}^QA θ_i(p) A_i to compute a lower-bound approximation to the minimum singular value of A.

It is additionally a functor as calling scm_init(p) on a parameter vector p will return the lower-bound estimate of the minimum singular value of A(p).

ModelOrderReductionToolkit.initialize_SCM_SPD — Function

initialize_SCM_SPD(param_disc,Ais,makeθAi,Mα,Mp,ϵ;[optimizer=Tulip.Optimizer,noise=1]) = SCM_Init

Method to initialize an SCM_Init object to perform the SCM on an affinely-parameter-dependent symmetric positive definite matrix A(p) = ∑ makeθAi(p,i) Ais[i] to compute a lower-bound approximation to the minimum singular value (or eigenvalue) of A(p).

Parameters:

param_disc: Either a matrix where each column is a parameter value in the discretization, or a vector of parameter vectors.

Ais: A vector of matrices, of the same dimension, used to construct the full matrix A(p) = ∑ makeθAi(p,i) Ais[i]

makeθAi(p,i): A function that takes in as input a parameter vector and an index such that A(p) = ∑ makeθAi(p,i) Ais[i]

Mα: Stability constraint constant (a positive integer)

Mp: Positivity constraint constant (a positive integer)

ϵ: Relative difference allowed between upper-bound and lower-bound approximation on the parameter discretization (between 0 and 1)

optimizer: Optimizer to pass into JuMP Model method for solving linear programs

noise: Determines amount of printed information, between 0 and 2 with 0 being nothing displayed; default 1

ModelOrderReductionToolkit.initialize_SCM_Noncoercive — Function

initialize_SCM_Noncoercive(param_disc,Ais,makeθAi,Mα,Mp,ϵ;[optimizer=Tulip.Optimizer,noise=1]) = SCM_Init

Method to initialize an SCM_Init object to perform the SCM on an affinely-parameter-dependent matrix A(p) = ∑ makeθAi(p,i) Ais[i] to compute a lower-bound approximation to the minimum singular value of A.

Parameters:

param_disc: Either a matrix where each column is a parameter value in the discretization, or a vector of parameter vectors.

Ais: A vector of matrices, of the same dimension, used to construct the full matrix A(p) = ∑ makeθAi(p,i) Ais[i]

makeθAi(p,i): A function that takes in as input a parameter vector and an index such that A(p) = ∑ makeθAi(p,i) Ais[i]

Mα: Stability constraint constant (a positive integer)

Mp: Positivity constraint constant (a positive integer)

ϵ: Relative difference allowed between upper-bound and lower-bound approximation on the parameter discretization (between 0 and 1)

optimizer: Optimizer to pass into JuMP Model method for solving linear programs

noise: Determines amount of printed information, between 0 and 2 with 0 being nothing displayed; default 1

ModelOrderReductionToolkit.find_sigma_bounds — Function

find_sigma_bounds(scm_init, p[, sigma_eps=1.0; noise=0])

Method that performs the online phase of SCM for the matrix A(p) = ∑ makeθAi(p,i) Ais[i] to compute lower and upper-bound approximations to the minimum singular value of A. Additional optional parameter sigma_eps such that if the computed ϵ difference of (σ_UB - σ_LB) / σ_UB is less than sigma_eps, we know that not enough stability constraints were enforced, and the minimum singular value is directly computed, appended to the scm_init's upper-bound set, and returned as both the lower and upper-bounds.

ModelOrderReductionToolkit.form_upperbound_set! — Function

form_upperbound_set!(scm_init; [noise=1])

Helper method that takes in an SCM_Init_SPD object and forms the upper-bound sets C, Y_UB, and σ_UBs until the ϵ tolerance provided in scm_init is met across the parameter discretization. The noise input determines the amount of printed information, between 0 and 2, with 0 being nothing displayed, and 1 is default.

ModelOrderReductionToolkit.solve_LBs_LP — Function

solve_LBs_LP(scm_init, p[; noise=1]) = (σ_LB, y_LB)

Helper method that takes in an SCM_Init_SPD object and a parameter vector p, and sets up and solves a linear program to compute a lower-bound σ_LB to the minimum singular value of A(p) along with the associated vector y_LB.

Radial-basis interpolatory stability factor

ModelOrderReductionToolkit.Sigma_Min_RBF — Type

Sigma_Min_RBF

A mutable struct used for approximating the minimum singular value of a matrix with parametric dependence A(p). Given a new parameter value p, approximate σ_min(A(p)) with sigma_min_rbf(p).

ModelOrderReductionToolkit.min_sigma_rbf — Function

min_sigma_rbf(params, Ais, makeθAi[, ϕ=gaussian_rbf])

Method to form a interpolatory radial-basis function functor to approximate the minimum singular value of a parametrized matrix A(p). Pass as input either a matrix params where each column is a parameter vector, or a vector of vectors params where each vector is a parameter.

Optional argument for the radial-basis function ϕ, which defaults to the Gaussian ϕ(r) = exp(-r^2).

Returns a functor sigma_min_rbf <: Sigma_Min_RBF such that given a new parameter vector p, sigma_min_rbf(p) returns an approximation to the minimum singular value of A(p).

Offline, solves for the minimum singular value for each parameter in params, and uses this to form an interpolatory approximation in the form of log(σ_min(A(p))) ≈ ω_0 + ∑ (ω_i p_i) + ∑ (γ_i ϕ(p - p_i)) where p_i is the i'th parameter in params, and ω_0, ω_i, and γ_i are determined by the given params such that the approximation holds true for all p_i, and that ∑ γ_i = 0, and that ∑ γ_i p_i[j] = 0 for each j.

ModelOrderReductionToolkit.update_sigma_rbf! — Function

update_sigma_rbf!(sigma_min_rbf, ϕ)

Given a Sigma_Min_RBF object sigma_min_rbf, update its RBF, ϕ, and update the coefficients.

Computation of norm of residual

ModelOrderReductionToolkit.Affine_Residual_Init — Type

Affine_Residual_Init

A struct for containing the necessary vectors and matrices for quickly compute the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Here, u solves A(p) u = b(p) with A and b having affine parameter dependence, and V is a matrix with columns defining bases for approximation spaces u ≈ V u_r.

ModelOrderReductionToolkit.residual_norm_affine_init — Function

residual_norm_affine_init(Ais, makeθAi, bis, makeθbi, V[, X0=nothing, T=Float64])

Method that constructs the necessary vectors and matrices to quickly compute the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Pass as input a vector of matrices Ais, and a function makeθAi such that the affine construction of A is given by A(p) = ∑_{i=1}^QA makeθAi(p,i) * Ais[i], and similarly a vector of vectors bis and a function makeθbi such that the affine construction of b is given by b(p) = ∑_{i=1}^Qb makeθbi(p,i) * bis[i].

Additionally, pass in a matrix V which contains as columns a basis for a reduced space, u ≈ V u_r with the dimension of u_r less than that of u.

Optionally pass in a matrix X from which the X-norm of the residual will be computed in the method residual_norm_affine_online. If X remains as nothing, then will choose it to be the identity matrix to compute the 2-norm of the residual.

If using complex numbers, specify T=ComplexF64.

ModelOrderReductionToolkit.add_col_to_V! — Function

add_col_to_V!(res_init, v, T)

Method to add a vector v to the columns of the matrix V in the Affine_Residual_Init object, res_init, without recomputing all terms. Must specify type T.

ModelOrderReductionToolkit.residual_norm_affine_online — Function

residual_norm_affine_online(res_init, u_r, p)

Method that given res_init, an Affine_Residual_Init object, computes the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Pass as input the Affine_Residual_Init object, res_init, a reduced vector u_r, and the corresponding parameter vector p.

ModelOrderReductionToolkit.residual_norm_explicit — Function

residual_norm_explicit(u_approx, p, makeA, makeb, X=nothing)

Method that computes the X-norm of the residual, r(u_r,p) = A(p) (u - u_approx) = b(p) - A(p) u_approx explicitly where u is the solution to A(p) u = b(p), and u_approx is an approximation, u ≈ u_approx.

Pass as input the approximation vector u_approx, the corresponding parameter, p, the method makeA(p) for constructing the matrix A(p), and the method makeb(p) for constructing the vector b(p).

Optionally pass in a matrix X from which the X-norm of the residual will be computed. If X remains as nothing, then will choose it to be the identity matrix to compute the 2-norm of the residual.

residual_norm_explicit(u_approx, p, Ais, makeθAi, bis, makeθbi, X=nothing)

Method that computes the X-norm of the residual, r(u_r,p) = A(p) (u - u_approx) = b(p) - A(p) u_approx explicitly where x is the solution to A(p) u = b(p), and u_approx is an approximation, u ≈ u_approx.

Pass as input the approximation vector u_approx, the corresponding parameter, p, a vector of matrices Ais, and a function makeθAi such that the affine construction of A is given by A(p) = ∑_{i=1}^QA makeθAi(p,i) * Ais[i], and similarly a vector of vectors bis and a function makeθbi such that the affine construction of b is given by b(p) = ∑_{i=1}^Qb makeθbi(p,i) * bis[i].

Optionally pass in a matrix X from which the X-norm of the residual will be computed. If X remains as nothing, then will choose it to be the identity matrix to compute the 2-norm of the residual.