Additional ModelOrderReductionToolkit.jl Docstrings

Models

ModelOrderReductionToolkit.galerkin_project — Function

rom = galerkin_project(model::LinearModel, V[, W=V; r=-1])

Perform (Petrov) Galerkin projection on a linear model where the trial space is the first r columns of V and the test space is the first r columns of W. If r=-1, uses all columns. Returns a new LinearModel.

W' * A(p) * V * x_r = W' * b(p),V * x_r ≈ x = A(p)^(-1) b(p)`

rom = galerkin_project(model::LinearMatrixModel, V[, W=V; r=-1])

Perform (Petrov) Galerkin projection on each linear model in model.models.

galerkin_project(model, V[, W=V, ])

ModelOrderReductionToolkit.galerkin_add! — Function

galerkin_add!(rom::LinearModel, fom::LinearModel, v, Vold[, w=v, Wold=Vold; r=-1])

Assuming that rom = galerkin_project(model, Vold, Wold), updates rom such that if V = [Vold v] and W = [Wold w], then rom = galerkin_project(model, V, W).

galerkin_add!(rom::LinearMatrixModel, fom::LinearMatrixModel, v, Vold[, w=v, Wold=Vold; r=-1])

Assuming that rom = galerkin_project(fom, Vold, Wold), updates rom such that if V = [Vold v] and W = [Wold w], then rom = galerkin_project(fom, V, W).

ModelOrderReductionToolkit.PoissonModel — Function

model = PoissonModel([; Nx=999, P=3])

Uses finite differences on the following PDE to generate a LinearModel with parameter dependence. Change discretization with Nx, and change number of parameters, and hence number of affine terms, with P.

- ∂_x (κ(x, p) ∂_x u(x, p)) = f(x)

u(0,p) = 0; u(1,p) = p[1]

κ(x,p) = (1.05 - (1/2)^(P-1)) + 0.5 p[2] sin(2πx) + 0.25 p[3] sin(4πx) + ... + (1/2)^(P-1) p[P] sin(2π * P * x)

f(x,p) = (0.25 .< x .< 0.75) .* 10.0

length(p) = P; 0 ≤ p[i] ≤ 1

ModelOrderReductionToolkit.PenzlModel — Function

model = PenzlModel()

Generates the standard Penzl LTIModel with one input, one output, and 1006 dimension state variable.

ModelOrderReductionToolkit.MISOPenzlModel — Function

model = MISOPenzlModel()

Generates an LTIModel with three inputs, one output, and a state of dimension 1006. Same structure as the Penzl model except the B matrix is changed to

1006×3 Matrix{Float64}:
 10.0   0.0   0.0
 10.0   0.0   0.0
  0.0  10.0   0.0
  0.0  10.0   0.0
  0.0   0.0  10.0
  0.0   0.0  10.0
  1.0   1.0   1.0
  ⋮
  1.0   1.0   1.0
  1.0   1.0   1.0
  1.0   1.0   1.0
  1.0   1.0   1.0
  1.0   1.0   1.0
  1.0   1.0   1.0
  1.0   1.0   1.0

ModelOrderReductionToolkit.to_frequency_domain — Function

frequency_model = to_frequency_domain(model)

Assuming null initial conditions, uses the Laplace transform to convert the LTIModel into a LinearMatrixModel for which the first element of the parameter vector is the imaginary part of the frequency variable. The model is of the form sE(p) X = A(p) X + B(p) where X is the Laplace variable for X and s = 0 + iω

ModelOrderReductionToolkit.to_ss — Function

to_ss(model[, p=nothing])

Initializes the model to the parameter p if passed in, then returns a ControlSystems.jl StateSpace object.

ModelOrderReductionToolkit.to_dss — Function

to_dss(model[, p=nothing])

Initializes the model to the parameter p if passed in, then returns a DescriptorSystems.jl DescriptorStateSpace object.

ModelOrderReductionToolkit.to_ode_problem — Function

to_ode_problem(model[, p=nothing; u=(t->zeros(size(model.B, 2))), x0=0.0, tspan=(0,1)])

Creates an ODEProblem for the model <: LTISystem for a given input u(t). Note that this is the ODE for the state variable x. Once have formed the solution object, will have to multiply by model.C to get the output y. Note that DifferentialEquations.jl names the output u, which for this problem is the state variable x, not the input u.

ModelOrderReductionToolkit.ParameterizedPenzlModel — Function

model = ParameterizedPenzlModel()

Generates an LTIModel with one input, one output, and a state of dimension 1006. Same structure as the Penzl model, expect depends on a parameter vector of length 3 which shift the poles along the complex axis. Instantiate to a parameter vector by calling model([p1,p2,p3]).

Reductors

ModelOrderReductionToolkit.StabilityResidualErrorEstimator — Type

StabilityResidualErrorEstimator <: ErrorEstimator

Error estimator for a LinearModel which approximates the stability factor with stability_estimator(p) and computes the residual norm with residual_norm.

ModelOrderReductionToolkit.form_rom — Function

form_rom(pod_reductor, r=-1)

Pulls the first r left singular vectors from pod_reductor.decomp, and then Galerkin projects pod_reductor.model onto that basis, and returns the resulting ROM.

form_rom(sg_reductor, r=-1)

Calls galerkin_project on the FOM and returns a ROM with RB of dimension r. If r=-1, uses all available columns of sg_reductor.V.

form_rom(wg_reductor, r=-1)

Calls galerkin_project on the FOM and returns a ROM with RB of dimension r. If r=-1, uses all available columns of wg_reductor.V.

ModelOrderReductionToolkit.lift — Function

lift(pod_reductor, x_r)

Given a vector solution x_r to a ROM formed by the pod_reductor, which is of smaller dimension than outputs of the FOM, lifts the solution to the same dimension of the FOM.

lift(sg_reductor, x_r)

Given a vector solution x_r to a ROM formed by the sg_reductor, which is of smaller dimension than outputs of the FOM, lifts the solution to the same dimension of the FOM.

lift(wg_reductor, x_r)

Given a vector solution x_r to a ROM formed by the wg_reductor, which is of smaller dimension than outputs of the FOM, lifts the solution to the same dimension of the FOM.

ModelOrderReductionToolkit.add_to_rb! — Function

add_to_rb!(pod_reductor, snapshots[; noise=1])

Directly updates pod_reductor with new snapshots given in the columns of the matrix snapshots.

add_to_rb!(pod_reductor, parameters[; noise=1, progress=true])

Loops through the vector of parameters, forms their full order solutions, adds them to pod_reductor.snapshots, and then updates the singular values and singular vectors in pod_reductor.S and pod_reductor.V.

add_to_rb!(sg_reductor, snapshots[; noise=1])

Directly updates sg_reductor with new snapshots given in the columns of the matrix snapshots.

add_to_rb!(sg_reductor, parameters[; noise=1, progress=true])

Loops through the vector of parameters, forms their full order solutions, adds them to sg_reductor.snapshots, and then updates the reduced basis in sg_reductor.V.

add_to_rb!(wg_reductor, params[; noise=1, eps=0.0, zero_tol=1e-15])

Loops through the vector of parameters params, computes the approximate estimator for each, selects the one with the highest error, and updates wg_reductor with the corresponding full order solution. Returns true if a vector is added to the RB, false otherwise.

add_to_rb!(wg_reductor, params, r[; noise=1, eps=0.0, zero_tol=1e-15])

Adds to wg_reductor at least r times by calling add_to_rb!(wg_reductor, params, noise=noise, eps=eps, zero_tol=zero_tol) several times. If all r are added, returns true, otherwise false.

ModelOrderReductionToolkit.get_rom — Function

get_rom(wg_reductor)

Helper method for getting the ROM from the wg_reductor object. Can otherwise obtain it through wg_reductor.rom.

Affinely parameter-dependent arrays:

ModelOrderReductionToolkit.APArray — Type

AffineParametrizedArray(arrays, makeθi[, precompθ=false])

Struct for containing an affine parametrized array A(p) of the form A(p) = ∑ makeθi(p,i) * arrays[i].

Arrays are stored in the vector arrays for quick recomputation for new parameter values. Each must be of the same size and dimension so that they can be broadcast summed.

If precompθ set to false (default):

The function makeθi(p,i) takes in a parameter object first, and second an index i=1,...,length(arrays), and returns a scalar

If precompθ set to true:

The function makeθi(p) takes in a parameter object, and returns a vector of each of the affine terms.

Given aparr <: AffineParametrizedArray, and a new parameter value p, can form the full array A(p) by calling aparr(p).

ModelOrderReductionToolkit.formArray! — Function

formArray!(aparr, arr, p)

Given an array arr with the same dimensions as the arrays in the APArray aparr, form A(p) and place its values in arr.

ModelOrderReductionToolkit.eim — Function

eim(arrFun, param_disc[, ϵ=1e-2; maxM=100, noise=1])

Method for constructing an APArray object from a non-affinely parameter dependent matrix arrFun(p) by empirical interpolation.

param_disc must be a matrix with columns as parameter vectors, or a vector with elements as parameters.

Loops over the given parameter discretization until a maximum ∞-norm error of ϵ is achieved over the entire discretization, or until the maximum number of parameter values are chosen, given by maxM.

noise dictates the amount of printed output. Set 0 for no output, 1 for some output, ≥2 for most.

Matrices as vector of vectors:

ModelOrderReductionToolkit.VOV — Type

VectorOfVectors{T} <: AbstractMatrix{T}

Type for defining matrices as vectors of vectors for quick insertion of rows and columns. Stores the columns as vectors in vecs and its size in size.

Construct an empty VOV of dimensions (nrows × ncols) where one of nrows or ncols must be zero:

VectorOfVectors(nrows=0, ncols=0, T=Float64)

Construct a VOV from a Julia vector of vectors:

VectorOfVectors(vecs::AbstractVector{<:AbstractVector{T}})

Construct a VOV from a matrix:

VectorOfVectors(M::AbstractMatrix{T})

ModelOrderReductionToolkit.addRow! — Function

addRow!(vov::VectorOfVectors)

Add a row to the vector of vectors by appending zero(T) to each vector.

addRow!(vov::VectorOfVectors, row::AbstractVector)

Add the vector row to the last column of vov.

ModelOrderReductionToolkit.removeRow! — Function

removeRow!(vov::VectorOfVectors)

Remove a row from the vector of vectors.

ModelOrderReductionToolkit.addCol! — Function

addCol!(vov::VectorOfVectors)

Add a column to the vector of vectors by appending zeros(T,nrow) to vov.vecs.

addCol!(vov::VectorOfVectors, col::AbstractVector)

Add the vector row to the last column of vov.

ModelOrderReductionToolkit.removeCol! — Function

removeCol!(vov::VectorOfVectors)

Remove a column from the vector of vectors.

Successive constraint method (SCM):

ModelOrderReductionToolkit.SCM_Init — Type

SCM_Init is a struct for holding all of the variables and methods necessary for running the successive constraint method on an affinely-parameter-dependent matrix A(p) = ∑_{i=1}^QA θ_i(p) A_i to compute a lower-bound approximation to the minimum singular value of A.

It is additionally a functor as calling scm_init(p) on a parameter vector p will return the lower-bound estimate of the minimum singular value of A(p).

ModelOrderReductionToolkit.initialize_SCM_SPD — Function

initialize_SCM_SPD(param_disc,Ap,Mα,Mp,ϵ[;optimizer=Tulip.Optimizer,kmaxiter=1000,noise=1,lp_attrs]) = SCM_Init

Method to initialize an SCM_Init object to perform the SCM on an affinely-parameter-dependent symmetric positive definite matrix Ap<:APArray to compute a lower-bound approximation to the minimum singular value (stability factor) of A(p).

Parameters:

param_disc: Either a matrix where each column is a parameter value in the discretization, or a vector of parameter vectors.

Ap: APArray object that stores the affine parameter dependence of the matrix A(p) = ∑ makeθAi(p,i) Ais[i]

Mα: Stability constraint constant (a positive integer)

Mp: Positivity constraint constant (a positive integer)

ϵ: Relative difference allowed between upper-bound and lower-bound approximation on the parameter discretization (between 0 and 1)

kmaxiter: Maximum number of iterations used in iterating eigensolver before defaulting to full-dense eigensolve.

noise: Determines amount of printed information, between 0 and 2 with 0 being nothing displayed; default 1.

lp_attrs: A dicitonary of attributes to set for the linear program with the given optimizer.

ModelOrderReductionToolkit.initialize_SCM_Noncoercive — Function

initialize_SCM_Noncoercive(param_disc,Ap,Mα,Mp,ϵ[;optimizer=Tulip.Optimizer,kmaxiter=1000,noise=1,lp_attrs]) = SCM_Init

Method to initialize an SCM_Init object to perform the SCM on an affinely-parameter-dependent matrix A(p) = ∑ makeθAi(p,i) Ais[i] to compute a lower-bound approximation to the minimum singular value (stability factor) of A(p).

Parameters:

param_disc: Either a matrix where each column is a parameter value in the discretization, or a vector of parameter vectors.

Ap: APArray object to hold information for affine parameter dependence of matrix A(p).

Mα: Stability constraint constant (a positive integer)

Mp: Positivity constraint constant (a positive integer)

ϵ: Relative difference allowed between upper-bound and lower-bound approximation on the parameter discretization (between 0 and 1)

kmaxiter: Maximum number of iterations used in iterating eigensolver before defaulting to full-dense eigensolve.

noise: Determines amount of printed information, between 0 and 2 with 0 being nothing displayed; default 1.

lp_attrs: A dicitonary of attributes to set for the linear program with the given optimizer.

ModelOrderReductionToolkit.find_sigma_bounds — Function

find_sigma_bounds(scm_init, p[, sigma_eps=1.0; noise=0])

Method that performs the online phase of SCM for the matrix A(p) = ∑ makeθAi(p,i) Ais[i] to compute lower and upper-bound approximations to the minimum singular value of A. Additional optional parameter sigma_eps such that if the computed ϵ difference of (σ_UB - σ_LB) / σ_UB is less than sigma_eps, we know that not enough stability constraints were enforced, and the minimum singular value is directly computed, appended to the scm_init's upper-bound set, and returned as both the lower and upper-bounds.

Radial-basis interpolatory stability factor

ModelOrderReductionToolkit.Sigma_Min_RBF — Type

Sigma_Min_RBF

A mutable struct used for approximating the minimum singular value of a matrix with parametric dependence A(p). Given a new parameter value p, approximate σ_min(A(p)) with sigma_min_rbf(p).

ModelOrderReductionToolkit.min_sigma_rbf — Function

min_sigma_rbf(params, Ais, makeθAi[, ϕ=gaussian_rbf])

Method to form a interpolatory radial-basis function functor to approximate the minimum singular value of a parametrized matrix A(p). Pass as input either a matrix params where each column is a parameter vector, or a vector of vectors params where each vector is a parameter.

Optional argument for the radial-basis function ϕ, which defaults to the Gaussian ϕ(r) = exp(-r^2).

Returns a functor sigma_min_rbf <: Sigma_Min_RBF such that given a new parameter vector p, sigma_min_rbf(p) returns an approximation to the minimum singular value of A(p).

Offline, solves for the minimum singular value for each parameter in params, and uses this to form an interpolatory approximation in the form of log(σ_min(A(p))) ≈ ω_0 + ∑ (ω_i p_i) + ∑ (γ_i ϕ(p - p_i)) where p_i is the i'th parameter in params, and ω_0, ω_i, and γ_i are determined by the given params such that the approximation holds true for all p_i, and that ∑ γ_i = 0, and that ∑ γ_i p_i[j] = 0 for each j.

ModelOrderReductionToolkit.update_sigma_rbf! — Function

update_sigma_rbf!(sigma_min_rbf, ϕ)

Given a Sigma_Min_RBF object sigma_min_rbf, update its RBF, ϕ, and update the coefficients.

Computation of norm of residual

ModelOrderReductionToolkit.ResidualNormComputer — Type

res <: ResidualNormComputer{T}

Abstract type for computing the norm of the residual for weak greedy RB methods. Must implement the update!(res, v) and compute(res, u_r, p) methods.

ModelOrderReductionToolkit.StandardResidualNormComputer — Type

StandardResidualNormComputer{T} <: ResidualNormComputer{T} StandardResidualNormComputer(Ap::APArray,bp::APArray,V=nothing,X=nothing)

A struct for containing the necessary vectors and matrices for quickly compute the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Here, u solves A(p) u = b(p) with A and b having affine parameter dependence, and V is a matrix with columns defining bases for approximation spaces u ≈ V u_r.

ModelOrderReductionToolkit.ProjectionResidualNormComputer — Type

ProjectionResidualNormComputer{T} <: ResidualNormComputer{T} ProjectionResidualNormComputer(Ap::APArray,bp::APArray,V=nothing,X=nothing)

A struct for containing the necessary vectors and matrices for quickly compute the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p). Uses a projection method which is more stable than the standard method.

Here, u solves A(p) u = b(p) with A and b having affine parameter dependence, and V is a matrix with columns defining bases for approximation spaces u ≈ V u_r.

Linear algebra utilities

ModelOrderReductionToolkit.full_lu — Function

full_lu(A; steps=-1)

Performs a completely pivoted LU factorization on the matrix A, returning permutation vectors Q and P, and lower and upper triangular matrices L and U, such that if all steps are performed, then A[P,Q] = L*U.

ModelOrderReductionToolkit.smallest_real_eigval — Function

smallest_real_eigval(A, kmaxiter[, noise=1, krylovsteps=10])

Given a hermitian matrix A, attempts to compute the most negative (real) eigenvalue. First, uses Krylov iteration with a shift-invert procedure with Gershgorin disks, and if not successful, calls a full, dense, eigensolve.

ModelOrderReductionToolkit.largest_real_eigval — Function

largest_real_eigval(A, kmaxiter[, noise=1])

Given a hermitian matrix A, attempts to compute the most positive (real) eigenvalue. First, uses Krylov iteration with no shift-invert, and if not successful, calls a full, dense, eigensolve.

ModelOrderReductionToolkit.smallest_real_pos_eigpair — Function

smallest_real_pos_eigpair(A, kmaxiter[, noise=1])

Given a hermitian, positive definite matrix A, attempts to compute the smallest (real) eigenvalue and eigenvector. First, uses Krylov iteration with shift-invert around zero, and if not successful, calls a full, dense, eigensolve. Returns a tuple with the first component being the eigenvalue, and the second component being the eigenvector.

ModelOrderReductionToolkit.smallest_sval — Function

smallest_sval(A, kmaxiter[, noise=1])

Given a matrix A, attempts to compute the smallest singular value of it by Krylov iteration and inversion around 0. If unsuccessful, computes a full, dense svd.

ModelOrderReductionToolkit.orthonormalize_mgs2! — Function

orthonormalize_mgs2!(u, V)

Given a matrix V, and a new vector u, orthogonalize u with respect to the columns of V, and computes its norm nu. If nu != 0, divides u by nu and returns nu. If nu == 0, then u lives in the span of V, and 0 is returned.