ModelOrderReductionToolkit.jl Docstrings

Affinely parameter-dependent arrays:

ModelOrderReductionToolkit.APArray — Type

AffineParametrizedArray(arrays, makeθi[, precompθ=false])

Struct for containing an affine parametrized array A(p) of the form A(p) = ∑ makeθi(p,i) * arrays[i].

Arrays are stored in the vector arrays for quick recomputation for new parameter values. Each must be of the same size and dimension so that they can be broadcast summed.

If precompθ set to false (default):

The function makeθi(p,i) takes in a parameter object first, and second an index i=1,...,length(arrays), and returns a scalar

If precompθ set to true:

The function makeθi(p) takes in a parameter object, and returns a vector of each of the affine terms.

Given aparr <: AffineParametrizedArray, and a new parameter value p, can form the full array A(p) by calling aparr(p).

ModelOrderReductionToolkit.formArray! — Function

formArray!(aparr, arr, p)

Given an array arr with the same dimensions as the arrays in the APArray aparr, form A(p) and place its values in arr.

ModelOrderReductionToolkit.eim — Function

eim(arrFun, param_disc[, ϵ=1e-2; maxM=100, noise=1])

Method for constructing an APArray object from a non-affinely parameter dependent matrix arrFun(p) by empirical interpolation.

param_disc must be a matrix with columns as parameter vectors, or a vector with elements as parameters.

Loops over the given parameter discretization until a maximum ∞-norm error of ϵ is achieved over the entire discretization, or until the maximum number of parameter values are chosen, given by maxM.

noise dictates the amount of printed output. Set 0 for no output, 1 for some output, ≥2 for most.

Matrices as vector of vectors:

ModelOrderReductionToolkit.VOV — Type

VectorOfVectors{T} <: AbstractMatrix{T}

Type for defining matrices as vectors of vectors for quick insertion of rows and columns. Stores the columns as vectors in vecs and its size in size.

Construct an empty VOV of dimensions (nrows × ncols) where one of nrows or ncols must be zero:

VectorOfVectors(nrows=0, ncols=0, T=Float64)

Construct a VOV from a Julia vector of vectors:

VectorOfVectors(vecs::AbstractVector{<:AbstractVector{T}})

Construct a VOV from a matrix:

VectorOfVectors(M::AbstractMatrix{T})

ModelOrderReductionToolkit.addRow! — Function

addRow!(vov::VectorOfVectors)

Add a row to the vector of vectors by appending zero(T) to each vector.

addRow!(vov::VectorOfVectors, row::AbstractVector)

Add the vector row to the last column of vov.

ModelOrderReductionToolkit.removeRow! — Function

removeRow!(vov::VectorOfVectors)

Remove a row from the vector of vectors.

ModelOrderReductionToolkit.addCol! — Function

addCol!(vov::VectorOfVectors)

Add a column to the vector of vectors by appending zeros(T,nrow) to vov.vecs.

addCol!(vov::VectorOfVectors, col::AbstractVector)

Add the vector row to the last column of vov.

ModelOrderReductionToolkit.removeCol! — Function

removeCol!(vov::VectorOfVectors)

Remove a column from the vector of vectors.

Projections functionality:

ModelOrderReductionToolkit.singular_values_information — Function

singular_values_information(σs)

Given a vector of real singular values σs, return a vector providing the maximum percentage of "information" available from a d-dimensional representation in Frobenius norm.

Uses the fact that the maximum error in projecting onto the first d left singular vectors is given by

||A-A_d||_F = √ (σ_{d+1}^2 + ... + σ_N^2).

So the information with a d-dimensional representation given by

singular_values_information[d] = 1 - √ (σ_{d+1}^2 + ... + σ_N^2) / √ (σ_1^2 + ... + σ_N^2)

ModelOrderReductionToolkit.pca_projector — Function

P = pca_projector(A, dim=-1; info_tol=1e-2)

Uses the first dim singular vectors of A to generate an orthogonal projection operator P. Operator performs matrix operation and can be done on vectors or matrices of proper dimension.

If dim is not provided, uses n singular values where n is the smallest integer such that the amount of information per the Frobenious norm is at least 1-info_tol.

Returns a functor P which stores the reduced dimension n, the dimension of vectors to project m, the singular values of A in s, and the vectors to project onto in the m × n matrix M.

Call with P(b) where b is a m vector or a matrix with m rows, projecting each column.

ModelOrderReductionToolkit.qr_projector — Function

qr_projector(A, dim)

Uses the first dim columns of A by qr-pivot order (norm) to generate an orthogonal projection operator P. Operator performs matrix operation and can be done on vectors or matrices of proper dimension.

Returns a functor P which stores the reduced dimension n, the dimension of vectors to project m, the vectors to project onto in the m × n matrix M, and the matrix inv(M'M) in MtMinv.

Call with P(b) where b is a m vector or a matrix with m rows, projecting each column.

ModelOrderReductionToolkit.eim_projector — Function

P = eim_projector(A, dim)

Uses discrete empirical interpolation to generate a projector operator P that selects a few rows from the given vector or matrix, and then projects them onto a set of columns chosen by a greedy procedure.

Returns a functor P which stores the reduced dimension n, the dimension of vectors to project m, the matrix A_C which is A with n extracted columns, and the matrix A_CR which is A with n extracted rows and columns.

Call with P(b) where b is a m vector or a matrix with m rows, projecting each column.

Greedy linear affine reduced basis

ModelOrderReductionToolkit.Greedy_RB_Affine_Linear — Type

Greedy_RB_Affine_Linear

Struct for containing initialized greedy reduced-basis method for parametrized problem A(p) u = b(p) with affine parameter dependence A(p) = ∑ makeθAi(p,i) Ais[i], and b(p) = ∑ makeθbi(p,i) bis[i].

Uses Galerkin projection onto span of columns of V, V' A(p) V u_r = V' b(p), with V u_r ≈ u = A(p)^(-1) b(p).

Given a new parameter vector, p, and an object greedy_sol::Greedy_RB_Affine_Linear, form the reduced basis solution with greedy_sol(p[, full=true]).

ModelOrderReductionToolkit.GreedyRBAffineLinear — Function

GreedyRBAffineLinear(param_disc, Ap, bp, approx_stability_factor, compute_res_norm, update_res_norm![, ϵ=1e-2; T=Float64, max_snapshots=-1, noise=1])

Constructs a Greedy_RB_Affine_Linear object for parametrized problem Ap(p) u(p) = bp(p) with affine parameter dependence given by APArrays Ap (matrix) and bp (vector)

param_disc must either be a matrix with columns as parameters, or a vector of parameters. Uses the function approx_stability_factor(p) to approximate the stability factor for each parameter in the discretization, and compute_res_norm(u_r,p) to approximate the norm of the residual. Utilizes the affine decomposition of A(p) and b(p) to compute upper-bounds for the error: ||u - V u_r|| <= ||b(p) - A(p) V u_r|| / σ_min(A(p)) = compute_res_norm(u_r,p) / approx_stability_factor(p). After the parameter value with the highest upperbound error is selected, its truth solution is computed, and it is added to the reduced basis. The method update_res_norm!(u) is used to update the residual norm computation.

This procedure loops until the upper-bound error is less than ϵ, or until the number of snapshots exceeds max_snapshots, or the minimum of number of parameters and number of columns of A by default.

Specify T if wish to use number type other than Float64, works for ComplexF64.

noise determines amount of printed output, 0 for no output, 1 for basic, and 2 for more.

GreedyRBAffineLinear(param_disc, Ap, bp, approx_stability_factor[, ϵ=1e-2; res_calc=2, T=Float64, max_snapshots=-1, noise=1])

Constructs a Greedy_RB_Affine_Linear object for parametrized problem Ap(p) u(p) = bp(p) with affine parameter dependence given by APArrays Ap (matrix) and bp (vector).

param_disc must either be a matrix with columns as parameters, or a vector of parameter vectors.

Uses res_calc to determine how the norm of the residual ||b(p) - A(p) V u_r|| is computed. By default set to 2 which uses a projection method and has online runtime independent of the dimension of A. If set to 1, uses a method that also has online runtime independent of the dimension of A, but may become numerically unstable for large reduced bases. If 0, computes the residual norm explicitly which scales like O(N) where N is the dimension of A.

Uses the function approx_stability_factor(p) to approximate the stability factor for each parameter in the discretization, and the above selection to compute the norm of the residual. Utilizes the affine decomposition of A(p) and b(p) to compute upper-bounds for the error: ||u - V u_r|| <= ||b(p) - A(p) V u_r|| / σ_min(A(p)) = compute_res_norm(u_r,p) / approx_stability_factor(p). After the parameter value with the highest upperbound error is selected, its truth solution is computed, and it is added to the reduced basis. The method update_res_norm!(u) is used to update the residual norm computation.

This procedure loops until the upper-bound error is less than ϵ, or until the number of snapshots exceeds max_snapshots, or the minimum of number of parameters and number of columns of A by default.

Specify T if wish to use number type other than Float64, works for ComplexF64.

noise determines amount of printed output, 0 for no output, 1 for basic, and 2 for more.

ModelOrderReductionToolkit.greedy_rb_err_data — Function

greedy_rb_err_data(param_disc,Ap,bp,approx_stability_factor[,num_snapshots=10,ϵ=1e-2;noise=1)

Constructs a Greedy_RB_Affine_Linear object for parametrized problem Ap(p) u(p) = bp(p) with affine parameter dependence given by APArrays Ap (matrix) and bp (vector).

Note: This method calls the full order solver on all parameters in param_disc, may take a long time to run.

See the docs for GreedyRBAffineLinear for more information.

Returns a dictionary, ret_data, with the following components:

ret_data[:truth_sols] = Vector containing all truth solutions.

ret_data[:basis_dim] - A vector of reduced basis dimensions from 2 to num_snapshots

ret_data[:weak_greedy_ub] - A vector of the maximum upperbound l2 error found by the (weak) greedy reduced basis method (see GreedyRBAffineLinear)

ret_data[:weak_greedy_true] - A vector of the truth l2 error of the vector chosen by the (weak) greedy reduced basis method (see GreedyRBAffineLinear)

ret_data[:weak_greedy_true_ub] - A vector of the maximum true l2 error found by the (weak) greedy reduced basis method (see GreedyRBAffineLinear)

ret_data[:strong_greedy_err] - A vector of the maximum l2 error found by a strong greedy reduced basis method, uses knowledge of all solutions

ret_data[:strong_greedy_proj] - A vector of the maximum l2 error found by projecting truth solutions onto the strong greedy reduced basis, uses knowledge of all solutions

ret_data[:qr_projector] - A projection object which can be used to orthogonally project vectors onto the strong greedy subspace. See ModelOrderReductionToolkit.qr_projector method.

ret_data[:pca_err] - A vector of the maximum l2 error found by a PCA/POD reduced basis method, uses knowledge of all solutions

ret_data[:strong_greedy_proj] - A vector of the maximum l2 error found by projecting truth solutions onto the PCA/POD reduced basis, uses knowledge of all solutions

ret_data[:pca_projector] - A projection object which can be used to orthogonally project vectors onto the SVD/PCA subspace. See ModelOrderReductionToolkit.pca_projector method.

ModelOrderReductionToolkit.append_affine_rbm! — Function

append_affine_rbm!(greedy_sol, x)

Given a Greedy_RB_Affine_Linear object, greedy_sol, and a new vector x, update the vectors and matrices in greedy_sol.

ModelOrderReductionToolkit.init_affine_rbm — Function

TODO Update header init_affine_rbm(x, Ais, bis, makeθAi, makeθbi, T)

Given a vector x, generates the matrices V and VtAVis, and the vectors Vtbis for preallocation and quick computation of reduced basis solutions for the problem A(p)x(p)=b(p) with affinely dependent matrix A(p) = ∑ makeθAi(p,i) Ais[i] and vector b(p) = ∑ makeθbi(p,i) bis[i]. Must pass in type T.

Successive constraint method (SCM):

ModelOrderReductionToolkit.SCM_Init — Type

SCM_Init is a struct for holding all of the variables and methods necessary for running the successive constraint method on an affinely-parameter-dependent matrix A(p) = ∑_{i=1}^QA θ_i(p) A_i to compute a lower-bound approximation to the minimum singular value of A.

It is additionally a functor as calling scm_init(p) on a parameter vector p will return the lower-bound estimate of the minimum singular value of A(p).

ModelOrderReductionToolkit.initialize_SCM_SPD — Function

initialize_SCM_SPD(param_disc,Ap,Mα,Mp,ϵ;[T=Float64,optimizer=Tulip.Optimizer,kmaxiter=1000,noise=1,lp_attrs]) = SCM_Init

Method to initialize an SCM_Init object to perform the SCM on an affinely-parameter-dependent symmetric positive definite matrix Ap<:APArray to compute a lower-bound approximation to the minimum singular value (stability factor) of A(p).

If wish to use complex matrices, pass in T=ComplexF64, however, affine coefficients must be real.

Parameters:

param_disc: Either a matrix where each column is a parameter value in the discretization, or a vector of parameter vectors.

Ap: APArray object that stores the affine parameter dependence of the matrix A(p) = ∑ makeθAi(p,i) Ais[i]

Mα: Stability constraint constant (a positive integer)

Mp: Positivity constraint constant (a positive integer)

ϵ: Relative difference allowed between upper-bound and lower-bound approximation on the parameter discretization (between 0 and 1)

T: Datatype for matrix initialization, default Float64. If using complex matrices, pass T=ComplexF64.

kmaxiter: Maximum number of iterations used in iterating eigensolver before defaulting to full-dense eigensolve.

noise: Determines amount of printed information, between 0 and 2 with 0 being nothing displayed; default 1.

lp_attrs: A dicitonary of attributes to set for the linear program with the given optimizer.

ModelOrderReductionToolkit.initialize_SCM_Noncoercive — Function

initialize_SCM_Noncoercive(param_disc,Ap,Mα,Mp,ϵ;[T=Float64,optimizer=Tulip.Optimizer,kmaxiter=1000,noise=1,lp_attrs]) = SCM_Init

Method to initialize an SCM_Init object to perform the SCM on an affinely-parameter-dependent matrix A(p) = ∑ makeθAi(p,i) Ais[i] to compute a lower-bound approximation to the minimum singular value (stability factor) of A(p).

If wish to use complex matrices, pass in T=ComplexF64, however, affine coefficients must be real.

Parameters:

param_disc: Either a matrix where each column is a parameter value in the discretization, or a vector of parameter vectors.

Ap: APArray object to hold information for affine parameter dependence of matrix A(p).

Mα: Stability constraint constant (a positive integer)

Mp: Positivity constraint constant (a positive integer)

ϵ: Relative difference allowed between upper-bound and lower-bound approximation on the parameter discretization (between 0 and 1)

T: Datatype for matrix initialization, default Float64. If using complex matrices, pass T=ComplexF64.

kmaxiter: Maximum number of iterations used in iterating eigensolver before defaulting to full-dense eigensolve.

noise: Determines amount of printed information, between 0 and 2 with 0 being nothing displayed; default 1.

lp_attrs: A dicitonary of attributes to set for the linear program with the given optimizer.

ModelOrderReductionToolkit.find_sigma_bounds — Function

find_sigma_bounds(scm_init, p[, sigma_eps=1.0; noise=0])

Method that performs the online phase of SCM for the matrix A(p) = ∑ makeθAi(p,i) Ais[i] to compute lower and upper-bound approximations to the minimum singular value of A. Additional optional parameter sigma_eps such that if the computed ϵ difference of (σ_UB - σ_LB) / σ_UB is less than sigma_eps, we know that not enough stability constraints were enforced, and the minimum singular value is directly computed, appended to the scm_init's upper-bound set, and returned as both the lower and upper-bounds.

Radial-basis interpolatory stability factor

ModelOrderReductionToolkit.Sigma_Min_RBF — Type

Sigma_Min_RBF

A mutable struct used for approximating the minimum singular value of a matrix with parametric dependence A(p). Given a new parameter value p, approximate σ_min(A(p)) with sigma_min_rbf(p).

ModelOrderReductionToolkit.min_sigma_rbf — Function

min_sigma_rbf(params, Ais, makeθAi[, ϕ=gaussian_rbf])

Method to form a interpolatory radial-basis function functor to approximate the minimum singular value of a parametrized matrix A(p). Pass as input either a matrix params where each column is a parameter vector, or a vector of vectors params where each vector is a parameter.

Optional argument for the radial-basis function ϕ, which defaults to the Gaussian ϕ(r) = exp(-r^2).

Returns a functor sigma_min_rbf <: Sigma_Min_RBF such that given a new parameter vector p, sigma_min_rbf(p) returns an approximation to the minimum singular value of A(p).

Offline, solves for the minimum singular value for each parameter in params, and uses this to form an interpolatory approximation in the form of log(σ_min(A(p))) ≈ ω_0 + ∑ (ω_i p_i) + ∑ (γ_i ϕ(p - p_i)) where p_i is the i'th parameter in params, and ω_0, ω_i, and γ_i are determined by the given params such that the approximation holds true for all p_i, and that ∑ γ_i = 0, and that ∑ γ_i p_i[j] = 0 for each j.

ModelOrderReductionToolkit.update_sigma_rbf! — Function

update_sigma_rbf!(sigma_min_rbf, ϕ)

Given a Sigma_Min_RBF object sigma_min_rbf, update its RBF, ϕ, and update the coefficients.

Computation of norm of residual

ModelOrderReductionToolkit.Affine_Residual_Init_Proj — Type

Affine_Residual_Init_Proj

A struct for containing the necessary vectors and matrices for quickly compute the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Here, u solves A(p) u = b(p) with A and b having affine parameter dependence, and V is a matrix with columns defining bases for approximation spaces u ≈ V u_r.

ModelOrderReductionToolkit.Affine_Residual_Init — Type

Affine_Residual_Init

A struct for containing the necessary vectors and matrices for quickly compute the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Here, u solves A(p) u = b(p) with A and b having affine parameter dependence, and V is a matrix with columns defining bases for approximation spaces u ≈ V u_r.

ModelOrderReductionToolkit.residual_norm_affine_proj_init — Function

residual_norm_affine_init(Ais, makeθAi, bis, makeθbi, V[, X=nothing, T=Float64])

Method that constructs the necessary vectors and matrices to quickly compute the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Does this in a more numerically stable way by computing the projections of b(p) onto the span of the columns of A(p) V, so no negative cross-term needs to be evaluated.

Pass as input a vector of matrices Ais, and a function makeθAi such that the affine construction of A is given by A(p) = ∑_{i=1}^QA makeθAi(p,i) * Ais[i], and similarly a vector of vectors bis and a function makeθbi such that the affine construction of b is given by b(p) = ∑_{i=1}^Qb makeθbi(p,i) * bis[i].

Additionally, pass in a matrix V which contains as columns a basis for a reduced space, u ≈ V u_r with the dimension of u_r less than that of u.

Optionally pass in a matrix X from which the X-norm of the residual will be computed in the method residual_norm_affine_online. If X remains as nothing, then will choose it to be the identity matrix to compute the 2-norm of the residual. (NOT YET IMPLEMENTED)

ModelOrderReductionToolkit.residual_norm_affine_init — Function

residual_norm_affine_init(Ais, makeθAi, bis, makeθbi, V[, X=nothing])

Method that constructs the necessary vectors and matrices to quickly compute the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Pass as input a vector of matrices Ais, and a function makeθAi such that the affine construction of A is given by A(p) = ∑_{i=1}^QA makeθAi(p,i) * Ais[i], and similarly a vector of vectors bis and a function makeθbi such that the affine construction of b is given by b(p) = ∑_{i=1}^Qb makeθbi(p,i) * bis[i].

Additionally, pass in a matrix V which contains as columns a basis for a reduced space, u ≈ V u_r with the dimension of u_r less than that of u.

Optionally pass in a matrix X from which the X-norm of the residual will be computed in the method residual_norm_affine_online. If X remains as nothing, then will choose it to be the identity matrix to compute the 2-norm of the residual. (NOT YET IMPLEMENTED)

ModelOrderReductionToolkit.add_col_to_V! — Function

add_col_to_V!(res_init, v)

Method to add a vector v to the columns of the matrix V in the Affine_Residual_Init object, res_init, without recomputing all terms.

add_col_to_V!(res_init, v)

Method to add a vector v to the columns of the matrix V in the Affine_Residual_Init_Proj object, res_init, without recomputing all terms.

ModelOrderReductionToolkit.residual_norm_affine_online — Function

residual_norm_affine_online(res_init, u_r, p)

Method that given res_init, an Affine_Residual_Init object, computes the X-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Pass as input the Affine_Residual_Init object, res_init, a reduced vector u_r, and the corresponding parameter vector p.

residual_norm_affine_online(res_init, u_r, p)

Method that given res_init, an Affine_Residual_Init_Proj object, computes the 2-norm of the residual, r(u_r,p) = A(p) (u - V u_r) = b(p) - A(p) V u_r, by taking advantage of affine parameter dependence of A(p) and b(p).

Does this in a more numerically stable way by computing the projections of b(p) onto the span of the columns of A(p) V, so no negative cross-term needs to be evaluated.

Pass as input the Affine_Residual_Init object, res_init, a reduced vector u_r, and the corresponding parameter vector p.

Linear algebra utilities

ModelOrderReductionToolkit.full_lu — Function

full_lu(A; steps=-1)

Performs a completely pivoted LU factorization on the matrix A, returning permutation vectors Q and P, and lower and upper triangular matrices L and U, such that if all steps are performed, then A[P,Q] = L*U.

ModelOrderReductionToolkit.smallest_real_eigval — Function

smallest_real_eigval(A, kmaxiter[, noise=1, krylovsteps=10])

Given a hermitian matrix A, attempts to compute the most negative (real) eigenvalue. First, uses Krylov iteration with a shift-invert procedure with Gershgorin disks, and if not successful, calls a full, dense, eigensolve.

ModelOrderReductionToolkit.largest_real_eigval — Function

largest_real_eigval(A, kmaxiter[, noise=1])

Given a hermitian matrix A, attempts to compute the most positive (real) eigenvalue. First, uses Krylov iteration with no shift-invert, and if not successful, calls a full, dense, eigensolve.

ModelOrderReductionToolkit.smallest_real_pos_eigpair — Function

smallest_real_pos_eigpair(A, kmaxiter[, noise=1])

Given a hermitian, positive definite matrix A, attempts to compute the smallest (real) eigenvalue and eigenvector. First, uses Krylov iteration with shift-invert around zero, and if not successful, calls a full, dense, eigensolve. Returns a tuple with the first component being the eigenvalue, and the second component being the eigenvector.

ModelOrderReductionToolkit.smallest_sval — Function

smallest_sval(A, kmaxiter[, noise=1])

Given a matrix A, attempts to compute the smallest singular value of it by Krylov iteration and inversion around 0. If unsuccessful, computes a full, dense svd.

ModelOrderReductionToolkit.orthonormalize_mgs2! — Function

orthonormalize_mgs2!(u, V)

Given a matrix V, and a new vector u, orthogonalize u with respect to the columns of V, and computes its norm nu. If nu != 0, divides u by nu and returns nu. If nu == 0, then u lives in the span of V, and 0 is returned.